working on the transfers enabled ante handler
